requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.9
django-allauth==65.13.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0
cryptography==46.0.3
